chore: split map logic from maps declaration
